Adding integration tests for endorsement service (#107)
* Added integration tests for endorsement service

* Added test suite for integration tests

* testing separation of unit and integration ci and pomxml

* Addressed review comments and added build step in CI

* Fix API contract and added pagination assertions to integration tests

* Fixing typos and addressing review comments

* Addressed review comments

* Addressed review comments

* Addressed review comments

---------
