+++ .github/workflows/ci.yml
@@ -6,7 +6,22 @@ on:
       - "**"
 jobs:
   build:
-    name: Maven Build
+     name: Maven Build
+     runs-on: ubuntu-latest
+     steps:
+       - uses: actions/checkout@v3
+
+       - name: Setup Java 17
+         uses: actions/setup-java@v3
+         with:
+           java-version: 17
+           distribution: "temurin"
+
+       - name: Build with Maven
+         run: mvn verify --file skill-tree/pom.xml -Pgit-build-profile -Dskip-tests=true
+
+  unit-test:
+    name: Unit Tests
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -15,7 +30,22 @@ jobs:
         uses: actions/setup-java@v3
         with:
           java-version: 17
-          distribution: 'temurin'
+          distribution: "temurin"
 
-      - name: Build with Maven
-        run: mvn verify --file skill-tree/pom.xml -Pgit-build-profile
+      - name: Unit Test
+        run: mvn test --file skill-tree/pom.xml
+
+  integration-test:
+    name: Integration Tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Setup Java 17
+        uses: actions/setup-java@v3
+        with:
+          java-version: 17
+          distribution: "temurin"
+
+      - name: Integration Test
+        run: mvn verify --file skill-tree/pom.xml -Dskip-ut=true
