+++ api-contracts.yml
@@ -170,7 +170,7 @@ paths:
           content:
             application/json:
               schema:
-                type: array
+                type: object
                 items:
                   $ref: '#/components/schemas/DummyEndorsementResponse'
