+++ skill-tree/pom.xml
@@ -196,6 +196,41 @@
                     </execution>
                 </executions>
             </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+
+                <configuration>
+                    <skip>${skip-ut}</skip>
+                    <skip>${skip-tests}</skip>
+                    <includes>
+                        <include>**/unit/*.java</include>
+                    </includes>
+
+                </configuration>
+
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <configuration>
+                    <skip>${skip-tests}</skip>
+                    <includes>
+                        <include>**/integration/*.java</include>
+                    </includes>
+
+                </configuration>
+
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>integration-test</goal>
+                            <goal>verify</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+
         </plugins>
     </build>
